refactor

+++ backend/src/routes/productRouter.js
@@ -1,23 +1,33 @@
 const KoaRouter = require('koa-router');
 const ProductController = require('../controllers/productController');
+const OrderController = require('../controllers/ordersController');
+const passport = require('../config/passport');
 
 const router = new KoaRouter();
 
 router
-  // Add new products
-  .post('/products', async (ctx, next) => {
-    await ProductController.create(ctx.request.body.product);
 
-    ctx.response.body = {message: 'Product created'}
+  .post('/products', async (ctx, next) => {
+    await passport.authenticate('jwt', {session: false},
+      async (err, user, msg) => {
+        if (err) throw new Error(err);
+
+        if (msg) throw new Error(msg.message);
+
+        if (user.type !== 3) {
+          throw new Error('Create a product allowed only admin');
+        }
+
+        await ProductController.create(ctx.request.body.product);
+        ctx.response.body = {message: 'Product created'};
+      })(ctx, next);
   })
 
-  // Get all available products
   .get('/products', async (ctx, next) => {
     const products = await ProductController.getAll();
     ctx.response.body = {products};
   })
 
-  // Get information about products
   .get('/products:id', async (ctx, next) => {
     const product =
       await ProductController.getById(ctx.params.id);
@@ -25,18 +35,41 @@ router
     ctx.response.body = {product};
   })
 
-  // Update products
   .put('/products:id', async (ctx, next) => {
-    await ProductController
-      .updateById(ctx.params.id, ctx.request.body);
+    await passport.authenticate('jwt', {session: false},
+      (err, user, msg) => {
+        if (err) throw new Error(err);
 
-    ctx.response.body = {message: 'Product updated'}
+        if (msg) throw new Error(msg.message);
+
+        if (user.type !== 3) {
+          throw new Error('Update product allowed only admin');
+        }
+
+        await ProductController
+          .updateById(ctx.params.id, ctx.request.body);
+
+        ctx.response.body = {message: 'Product updated'};
+      }
+    )(ctx, next);
   })
 
-  // Remove product
   .delete('/products:id', async (ctx, next) => {
-    await ProductController.deleteById(ctx.params.id);
-    ctx.response.body = {message: 'Product deleted'};
+    await passport.authenticate('jwt', {session: false},
+      (err, user, msg) => {
+        if (err) throw new Error(err);
+
+        if (msg) throw new Error(msg.message);
+
+        if (user.type !== 3) {
+          throw new Error('Delete product allowed only admin');
+        }
+
+        await ProductController.deleteById(ctx.params.id);
+        await OrderController.deleteByProductId(ctx.params.id);
+        ctx.response.body = {message: 'Product deleted'};
+      }
+    )(ctx, next);
   })
 
 module.exports = router;
